Move Sentry to ~/core/sentry.js
