+++ tests/common.js
@@ -4,7 +4,7 @@ var formatter = require('./lib/formatter')
 var Persistence = require('persistence')
 var RadarServer = require('../index.js').server
 var configuration = require('../configurator').load({persistence: true})
-var Sentry = require('../core/lib/resources/presence/sentry')
+var Sentry = require('../server/sentry')
 var Client = require('radar_client').constructor
 var fork = require('child_process').fork
 var Tracker = require('callback_tracker')
