+++ tests/lib/assert_helper.js
@@ -1,7 +1,7 @@
 var _ = require('underscore')
 var assert = require('assert')
 var EE = require('events').EventEmitter
-var Sentry = require('../../core/lib/resources/presence/sentry.js')
+var Sentry = require('../../server/sentry')
 var PresenceManager = require('../../core/lib/resources/presence/presence_manager.js')
 var SentryDefaults = {
   expiryOffset: 4000,
